requirements.txt
anyio==4.12.0
asgiref==3.11.0
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
Django==6.0
django-browser-reload==1.21.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
idna==3.11
msgpack==1.1.2
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
sqlparse==0.5.5
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
whitenoise==6.11.0
